Extract angle indicator rendering into reusable functions
